style(string): 🎨 | change count parameter type in repeat_string function from size_t to uint

+++ include/server/utils/string.h
@@ -9,11 +9,12 @@
 #ifndef STRING_H_
     #define STRING_H_
     #include <stddef.h>
+    #include <sys/types.h>
 
 void to_lowercase(char *str);
 void trim(char *str);
 char *dyn_strcat(char **original_dest, const char *src);
 char *empty_string(size_t size);
-char *repeat_string(const char *str, const char *delim, size_t count);
+char *repeat_string(const char *str, const char *delim, uint count);
 
 #endif /* !STRING_H_ */
